Expose ValidationHITCreator in the API
