+++ lib/index.js
@@ -12,6 +12,7 @@
 const { BasicSentenceGenerator, ContextualSentenceGenerator } = require('./sentence-generator');
 const SentenceSampler = require('./sampler');
 const { ParaphraseValidator, ParaphraseValidatorFilter } = require('./validator');
+const { ValidationHITCreator } = require('./validation');
 
 const DatasetAugmenter = require('./dataset_augmenter');
 const { DatasetParser, DatasetStringifier } = require('./dataset-parsers');
@@ -31,6 +32,7 @@ module.exports = {
     SentenceSampler,
     ParaphraseValidator,
     ParaphraseValidatorFilter,
+    ValidationHITCreator,
 
     BinaryPPDB,
